refactor: CarGameController 비즈니스 로직 제거
