+++ src/main/java/racingcargame/model/RacingCarGame.java
@@ -2,22 +2,30 @@ package racingcargame.model;
 
 import java.util.Collections;
 import java.util.List;
-import java.util.Objects;
 import java.util.stream.Collectors;
 
 public class RacingCarGame {
+    private static final String CAR_NAME_DELIMITER = ",";
     private static final int WINNER_POSITION = 0;
 
     private final RaceCount raceCount;
     private Cars carRepository;
 
-    public RacingCarGame(List<String> carNames, int raceCount) {
-        this.raceCount = Objects.requireNonNull(orderToSaveRaceCount(raceCount));
-        orderToSaveCars(carNames);
+    public RacingCarGame(String carNames, String raceCount) {
+        this.raceCount = orderToSaveRaceCount(makeIntegerRaceCount(raceCount));
+        this.carRepository = orderToSaveCars(makeSplitCarNames(carNames));
     }
 
-    private void orderToSaveCars(List<String> carNames) {
-        carRepository = new Cars(carNames);
+    private List<String> makeSplitCarNames(String carNames) {
+        return List.of(carNames.split(CAR_NAME_DELIMITER));
+    }
+
+    private int makeIntegerRaceCount(String count) {
+        return Integer.parseInt(count);
+    }
+
+    private Cars orderToSaveCars(List<String> carNames) {
+        return new Cars(carNames);
     }
 
     private RaceCount orderToSaveRaceCount(int count) {
